feat: Add date range iterator and monthly recurrence generator
This commit adds a new file `date-range.ts` which contains a generator function `iterate` that generates a range of dates between two given dates. It also adds a new file `monthly.ts` which contains a function `generate` that generates a list of dates based on a monthly recurrence rule.

The `iterate` function takes two parameters `from` and `to` which represent the start and end dates of the range. It uses a while loop to iterate over the dates and yield each date as a new instance of `Date`. The loop increments the date by one day using `setUTCDate` method.

The `generate` function takes three parameters `rule`, `from`, and `to`. The `rule` parameter is an object that contains a `bymonth` property which is an array of months. The function iterates over the dates generated by the `iterate` function and checks if the month of each date is included in the `bymonth` array. If it is, the date is added to the result array.

These changes are made to support the implementation of recurring events in the scheduler module.