feat: Refactor date range and recurrence functions
Refactor the code related to date range and recurrence functions. Move the `iterate` function from `date-range.ts` to `dates.ts` and rename it to `generateDateRange`. Also, move the `generate` function from `monthly.ts` to `dates.ts` and rename it to `generateMonthlyRecurrence`. These changes improve code organization and make it easier to maintain and reuse the functions.